feat(circle): break down `access.ts`
